fix: font poppins 적용 안되는 오류 수정
