+++ components/Typography/Headline/index.tsx
@@ -1,4 +1,13 @@
 import cn from 'classnames';
+import { Poppins } from 'next/font/google';
+
+const poppins = Poppins({
+  weight: ['500', '700'],
+  style: ['normal'],
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-poppins',
+});
 
 type HeadlineType = 'h1' | 'h2' | 'h3' | 'h4' | 'h5';
 
@@ -6,9 +15,9 @@ export type HeadlineProps = {
   className?: string;
   children: React.ReactNode;
   type: HeadlineType;
-  color?: keyof typeof TYPOGRAPHY_COLORS;  
+  color?: keyof typeof TYPOGRAPHY_COLORS;
   font?: 'Poppins' | 'Pretendard';
-  weights? : 400 | 500 | 600 | 700;
+  weights?: 400 | 500 | 600 | 700;
 };
 
 const fontSizes = {
@@ -26,16 +35,16 @@ const fontWeights = {
   700: 'font-bold',
 } as const;
 
-const Headline = ({ children, type, className, color, font, weights =700, ...rest }: HeadlineProps) => {
+const Headline = ({ children, type, className, color, font, weights = 700, ...rest }: HeadlineProps) => {
   const Component = type as keyof JSX.IntrinsicElements;
   const colorClass = color ? `text-${color}` : 'secondary-900';
 
-  const fontClass = font === 'Poppins' ? 'font-poppins': 'font-pretendard';
-  
+  const fontClass = font === 'Poppins' ? poppins.className : 'font-pretendard';
+
   const headlineStyle = `${fontClass} ${fontWeights[weights]} ${fontSizes[type]} leading-150 tracking-tighter ${colorClass}`;
 
   return (
-    <Component className= {cn(headlineStyle, className)} {...rest}>
+    <Component className={cn(headlineStyle, className)} {...rest}>
       {children}
     </Component>
   );
